finaly

+++ src/services/contacts.js
@@ -100,7 +100,7 @@ export const patchContact = async (id, payload, options = {}) => {
 
 export const deleteContact = async (id) => {
     try {
-        const contactToDelete = await contactsCollection.findOneAndDelete({ _id: id });
+        const contactToDelete = await contactsCollection.findOneAndDelete({ userId: id.userId, _id: id.contactId });
 
         return contactToDelete;
     } catch (error) {
